from czz, from oldsmth
git-svn-id: http://svn.kcn.cn/repos/kbs/trunk@9705 de4889a5-8ad4-4bf3-bd87-57997c21d515

+++ innbbsd/bbsnnrp.c
@@ -976,7 +976,7 @@ ULONG low, high;
 }
 
 int NNRPxhdrget(artno, mid, iscontrol)
-int *artno;
+ULONG *artno;
 char **mid;
 int iscontrol;
 {
